+++ src/components.tsx
@@ -1,16 +1,9 @@
 /*@refresh skip*/
+"use client";
 
-import type { Component, JSX } from "solid-js";
-import { children, createMemo, createRoot, mergeProps, on, Show, splitProps } from "solid-js";
-import { isServer, getRequestEvent } from "solid-js/web";
-import { pathIntegration, staticIntegration } from "./integration";
+import type { JSX } from "solid-js";
+import { createMemo, mergeProps, splitProps } from "solid-js";
 import {
-  createBranches,
-  createRouteContext,
-  createRouterContext,
-  getRouteMatches,
-  RouteContextObj,
-  RouterContextObj,
   useHref,
   useLocation,
   useNavigate,
@@ -18,19 +11,9 @@ import {
 } from "./routing";
 import type {
   Location,
-  LocationChangeSignal,
-  MatchFilters,
-  Navigator,
-  Params,
-  RouteContext,
-  RouteLoadFunc,
-  RouteDefinition,
-  RouterIntegration,
-  RouterContext,
-  Branch,
-  RouteSectionProps
+  Navigator
 } from "./types";
-import { normalizePath, createMemoObject } from "./utils";
+import { normalizePath } from "./utils";
 
 declare module "solid-js" {
   namespace JSX {
@@ -42,142 +25,6 @@ declare module "solid-js" {
     }
   }
 }
-
-export type RouterProps = {
-  base?: string;
-  actionBase?: string;
-  root?: Component<RouteSectionProps>;
-  children: JSX.Element;
-} & (
-  | {
-      url?: never;
-      source?: RouterIntegration | LocationChangeSignal;
-    }
-  | {
-      source?: never;
-      url: string;
-    }
-);
-
-export const Router = (props: RouterProps) => {
-  let e: any;
-  const { source, url, base, actionBase } = props;
-  const integration =
-    source ||
-    (isServer
-      ? staticIntegration({
-          value: url || ((e = getRequestEvent()) && getPath(e.request.url)) || ""
-        })
-      : pathIntegration());
-
-  const routeDefs = children(() => props.children) as unknown as () =>
-    | RouteDefinition
-    | RouteDefinition[];
-
-  const branches = createMemo(() =>
-    createBranches(
-      props.root ? { component: props.root, children: routeDefs() } : routeDefs(),
-      props.base || ""
-    )
-  );
-  const routerState = createRouterContext(integration, branches, { base, actionBase });
-
-  return (
-    <RouterContextObj.Provider value={routerState}>
-      <Routes routerState={routerState} branches={branches()} />
-    </RouterContextObj.Provider>
-  );
-};
-
-function getPath(url: string) {
-  const u = new URL(url);
-  return u.pathname + u.search;
-}
-
-function Routes(props: { routerState: RouterContext; branches: Branch[] }) {
-  const matches = createMemo(() =>
-    getRouteMatches(props.branches, props.routerState.location.pathname)
-  );
-  const params = createMemoObject(() => {
-    const m = matches();
-    const params: Params = {};
-    for (let i = 0; i < m.length; i++) {
-      Object.assign(params, m[i].params);
-    }
-    return params;
-  });
-  const disposers: (() => void)[] = [];
-  let root: RouteContext | undefined;
-
-  const routeStates = createMemo(
-    on(matches, (nextMatches, prevMatches, prev: RouteContext[] | undefined) => {
-      let equal = prevMatches && nextMatches.length === prevMatches.length;
-      const next: RouteContext[] = [];
-      for (let i = 0, len = nextMatches.length; i < len; i++) {
-        const prevMatch = prevMatches && prevMatches[i];
-        const nextMatch = nextMatches[i];
-
-        if (prev && prevMatch && nextMatch.route.key === prevMatch.route.key) {
-          next[i] = prev[i];
-        } else {
-          equal = false;
-          if (disposers[i]) {
-            disposers[i]();
-          }
-
-          createRoot(dispose => {
-            disposers[i] = dispose;
-            next[i] = createRouteContext(
-              props.routerState,
-              next[i - 1] || props.routerState.base,
-              createOutlet(() => routeStates()[i + 1]),
-              () => matches()[i],
-              params
-            );
-          });
-        }
-      }
-
-      disposers.splice(nextMatches.length).forEach(dispose => dispose());
-
-      if (prev && equal) {
-        return prev;
-      }
-      root = next[0];
-      return next;
-    })
-  );
-  return (
-    <Show when={routeStates() && root} keyed>
-      {route => <RouteContextObj.Provider value={route}>{route.outlet()}</RouteContextObj.Provider>}
-    </Show>
-  );
-}
-
-const createOutlet = (child: () => RouteContext | undefined) => {
-  return () => (
-    <Show when={child()} keyed>
-      {child => <RouteContextObj.Provider value={child}>{child.outlet()}</RouteContextObj.Provider>}
-    </Show>
-  );
-};
-
-export type RouteProps<S extends string> = {
-  path?: S | S[];
-  children?: JSX.Element;
-  load?: RouteLoadFunc;
-  matchFilters?: MatchFilters<S>;
-  component?: Component;
-};
-
-export const Route = <S extends string>(props: RouteProps<S>) => {
-  const childRoutes = children(() => props.children);
-  return mergeProps(props, {
-    get children() {
-      return childRoutes();
-    }
-  }) as unknown as JSX.Element;
-};
 
 export interface AnchorProps extends Omit<JSX.AnchorHTMLAttributes<HTMLAnchorElement>, "state"> {
   href: string;
@@ -224,8 +71,7 @@ export function A(props: AnchorProps) {
     />
   );
 }
-// deprecated alias exports
-export { A as Link, A as NavLink, AnchorProps as LinkProps, AnchorProps as NavLinkProps };
+
 export interface NavigateProps {
   href: ((args: { navigate: Navigator; location: Location }) => string) | string;
   state?: unknown;
