+++ src/data/events.ts
@@ -1,0 +1,129 @@
+import { delegateEvents } from "solid-js/web";
+import { onCleanup } from "solid-js";
+import type { RouterContext } from "../types";
+import { actions } from "./action";
+
+export function setupNativeEvents(router: RouterContext) {
+  const basePath = router.base.path();
+  const navigateFromRoute = router.navigatorFactory(router.base);
+  let preloadTimeout: Record<string, number> = {};
+
+  function isSvg<T extends SVGElement>(el: T | HTMLElement): el is T {
+    return el.namespaceURI === "http://www.w3.org/2000/svg";
+  }
+
+  function handleAnchor(evt: MouseEvent) {
+    if (
+      evt.defaultPrevented ||
+      evt.button !== 0 ||
+      evt.metaKey ||
+      evt.altKey ||
+      evt.ctrlKey ||
+      evt.shiftKey
+    )
+      return;
+
+    const a = evt
+      .composedPath()
+      .find(el => el instanceof Node && el.nodeName.toUpperCase() === "A") as
+      | HTMLAnchorElement
+      | SVGAElement
+      | undefined;
+
+    if (!a) return;
+
+    const svg = isSvg(a);
+    const href = svg ? a.href.baseVal : a.href;
+    const target = svg ? a.target.baseVal : a.target;
+    if (target || (!href && !a.hasAttribute("state"))) return;
+
+    const rel = (a.getAttribute("rel") || "").split(/\s+/);
+    if (a.hasAttribute("download") || (rel && rel.includes("external"))) return;
+
+    const url = svg ? new URL(href, document.baseURI) : new URL(href);
+    if (
+      url.origin !== window.location.origin ||
+      (basePath && url.pathname && !url.pathname.toLowerCase().startsWith(basePath.toLowerCase()))
+    )
+      return;
+    return [a, url] as const;
+  }
+
+  function handleAnchorClick(evt: Event) {
+    const res = handleAnchor(evt as MouseEvent);
+    if (!res) return;
+    const [a, url] = res;
+    const to = router.parsePath(url.pathname + url.search + url.hash);
+    const state = a.getAttribute("state");
+
+    evt.preventDefault();
+    navigateFromRoute(to, {
+      resolve: false,
+      replace: a.hasAttribute("replace"),
+      scroll: !a.hasAttribute("noscroll"),
+      state: state && JSON.parse(state)
+    });
+  }
+
+  function handleAnchorPreload(evt: Event) {
+    const res = handleAnchor(evt as MouseEvent);
+    if (!res) return;
+    const [a, url] = res;
+    if (!preloadTimeout[url.pathname]) router.preloadRoute(url, a.getAttribute("preload") !== "false");
+  }
+
+  function handleAnchorIn(evt: Event) {
+    const res = handleAnchor(evt as MouseEvent);
+    if (!res) return;
+    const [a, url] = res;
+    if (preloadTimeout[url.pathname]) return;
+    preloadTimeout[url.pathname] = setTimeout(() => {
+      router.preloadRoute(url, a.getAttribute("preload") !== "false");
+      delete preloadTimeout[url.pathname];
+    }, 200) as any;
+  }
+
+  function handleAnchorOut(evt: Event) {
+    const res = handleAnchor(evt as MouseEvent);
+    if (!res) return;
+    const [, url] = res;
+    if (preloadTimeout[url.pathname]) {
+      clearTimeout(preloadTimeout[url.pathname]);
+      delete preloadTimeout[url.pathname];
+    }
+  }
+
+  function handleFormSubmit(evt: SubmitEvent) {
+    let actionRef =
+      (evt.submitter && evt.submitter.getAttribute("formaction")) || (evt.target as any).action;
+    if (!actionRef) return;
+    if (!actionRef.startsWith("action:")) {
+      const url = new URL(actionRef);
+      actionRef = router.parsePath(url.pathname + url.search);
+      if (!actionRef.startsWith(router.actionBase)) return;
+    }
+    const handler = actions.get(actionRef);
+    if (handler) {
+      evt.preventDefault();
+      const data = new FormData(evt.target as HTMLFormElement);
+      handler.call(router, data);
+    }
+  }
+
+  // ensure delegated event run first
+  delegateEvents(["click", "submit"]);
+  document.addEventListener("click", handleAnchorClick);
+  document.addEventListener("mouseover", handleAnchorIn);
+  document.addEventListener("mouseout", handleAnchorOut);
+  document.addEventListener("focusin", handleAnchorPreload);
+  document.addEventListener("touchstart", handleAnchorPreload);
+  document.addEventListener("submit", handleFormSubmit);
+  onCleanup(() => {
+    document.removeEventListener("click", handleAnchorClick);
+    document.removeEventListener("mouseover", handleAnchorIn);
+    document.removeEventListener("mouseout", handleAnchorOut);
+    document.removeEventListener("focusin", handleAnchorPreload);
+    document.removeEventListener("touchstart", handleAnchorPreload);
+    document.removeEventListener("submit", handleFormSubmit);
+  });
+}
