+++ test/integration.spec.ts
@@ -1,4 +1,4 @@
-import { hashIntegration } from "../src/integration";
+import { hashParser } from "../src/routers/HashRouter";
 
 describe("Hash integration should", () => {
   test.each([
@@ -10,18 +10,16 @@ describe("Hash integration should", () => {
     ["file:///C:/Users/Foo/index.html#/test", "/test"]
   ])(`parse paths (case '%s' as '%s')`, (urlString, expected) => {
     const url = new URL(urlString);
-    const itegration = hashIntegration();
     const path = url.pathname + url.search + url.hash;
-    const parsed = itegration.utils!.parsePath!(path);
+    const parsed = hashParser(path);
     expect(parsed).toBe(expected);
   });
 
   test("parse hash-only paths to support in-page anchors", () => {
     window.location.hash = "/some-base-path";
     const url = new URL("http://localhost/#some-id");
-    const itegration = hashIntegration();
     const path = url.pathname + url.search + url.hash;
-    const parsed = itegration.utils!.parsePath!(path);
+    const parsed = hashParser(path);
     expect(parsed).toBe("/some-base-path#some-id");
   });
 });
