+++ test/router.spec.ts
@@ -10,7 +10,7 @@ describe("Router should", () => {
     test(`have a default path when base path is not defined`, () => {
       createRoot(() => {
         const signal = createSignal<LocationChange>({ value: "" });
-        const { base } = createRouterContext(signal);
+        const { base } = createRouterContext({ signal });
         expect(base.path()).toBe("/");
       });
     });
@@ -18,7 +18,7 @@ describe("Router should", () => {
     test(`have a normalized version of the base path when defined`, () => {
       createRoot(() => {
         const signal = createSignal<LocationChange>({ value: "" });
-        const { base } = createRouterContext(signal, fakeBranches, { base: "base" });
+        const { base } = createRouterContext({ signal }, fakeBranches, { base: "base" });
         expect(base.path()).toBe("/base");
       });
     });
@@ -26,7 +26,7 @@ describe("Router should", () => {
     test(`throw when the base path is invalid`, () => {
       createRoot(() => {
         const signal = createSignal<LocationChange>({ value: "" });
-        expect(() => createRouterContext(signal, fakeBranches, { base: "http://example.com" })).toThrow();
+        expect(() => createRouterContext({ signal }, fakeBranches, { base: "http://example.com" })).toThrow();
       });
     });
   });
@@ -37,7 +37,7 @@ describe("Router should", () => {
         const signal = createSignal<LocationChange>({
           value: "/foo/bar?hello=world"
         });
-        const { location } = createRouterContext(signal);
+        const { location } = createRouterContext({ signal });
         expect(location.pathname).toBe("/foo/bar");
         expect(location.search).toBe("?hello=world");
       });
@@ -50,7 +50,7 @@ describe("Router should", () => {
           const signal = createSignal<LocationChange>({
             value: "/foo/bar?hello=world"
           });
-          const { location } = createRouterContext(signal);
+          const { location } = createRouterContext({ signal });
           expect(location.pathname).toBe("/foo/bar");
           signal[1]({ value: expected + "?hello=world" });
           waitFor(() => signal[0]().value === expected + "?hello=world").then(() => {
@@ -64,7 +64,7 @@ describe("Router should", () => {
           const signal = createSignal<LocationChange>({
             value: "/foo/bar?hello=world"
           });
-          const { location } = createRouterContext(signal);
+          const { location } = createRouterContext({ signal });
           const count = createCounter(() => location.pathname);
           expect(location.pathname).toBe("/foo/bar");
           signal[1]({ value: "/foo/bar?fizz=buzz" });
@@ -77,7 +77,7 @@ describe("Router should", () => {
           const signal = createSignal<LocationChange>({
             value: "/foo bar+baz"
           });
-          const { location } = createRouterContext(signal);
+          const { location } = createRouterContext({ signal });
           expect(location.pathname).toBe("/foo%20bar+baz");
         }));
     }); // end of "contain property 'pathname'"
@@ -89,7 +89,7 @@ describe("Router should", () => {
           const signal = createSignal<LocationChange>({
             value: "/foo/bar?hello=world"
           });
-          const { location } = createRouterContext(signal);
+          const { location } = createRouterContext({ signal });
 
           expect(location.search).toBe("?hello=world");
           signal[1]({ value: "/foo/baz" + expected });
@@ -105,7 +105,7 @@ describe("Router should", () => {
           const signal = createSignal<LocationChange>({
             value: "/foo/bar?hello=world"
           });
-          const { location } = createRouterContext(signal);
+          const { location } = createRouterContext({ signal });
           const count = createCounter(() => location.search);
 
           expect(location.search).toBe("?hello=world");
@@ -119,7 +119,7 @@ describe("Router should", () => {
           const signal = createSignal<LocationChange>({
             value: "/foo?hello+world=bar+baz"
           });
-          const { location } = createRouterContext(signal);
+          const { location } = createRouterContext({ signal });
           expect(location.search).toBe("?hello+world=bar+baz");
         }));
     }); //end of "contain property 'search'"
@@ -130,7 +130,7 @@ describe("Router should", () => {
           const signal = createSignal<LocationChange>({
             value: "/foo#bar baz"
           });
-          const { location } = createRouterContext(signal);
+          const { location } = createRouterContext({ signal });
           expect(location.hash).toBe("#bar%20baz");
         }));
     }); // end of "contain property 'hash'"
@@ -141,7 +141,7 @@ describe("Router should", () => {
           const signal = createSignal<LocationChange>({
             value: "/foo/bar?hello=world&fizz=buzz"
           });
-          const { location } = createRouterContext(signal);
+          const { location } = createRouterContext({ signal });
           expect(location.query.hello).toEqual("world");
           expect(location.query.fizz).toEqual("buzz");
         });
@@ -152,7 +152,7 @@ describe("Router should", () => {
           const signal = createSignal<LocationChange>({
             value: "/foo/bar?hello=world"
           });
-          const { location } = createRouterContext(signal);
+          const { location } = createRouterContext({ signal });
 
           expect(location.query.hello).toEqual("world");
           signal[1]({ value: "/foo/bar?hello=world&fizz=buzz" });
@@ -168,7 +168,7 @@ describe("Router should", () => {
           const signal = createSignal<LocationChange>({
             value: "/foo/bar?hello=world"
           });
-          const { location } = createRouterContext(signal);
+          const { location } = createRouterContext({ signal });
           const count = createCounter(() => location.query.hello);
 
           expect(location.query.hello).toEqual("world");
@@ -188,7 +188,7 @@ describe("Router should", () => {
           const signal = createSignal<LocationChange>({
             value: "/foo/bar?hello=world"
           });
-          const { location } = createRouterContext(signal);
+          const { location } = createRouterContext({ signal });
           const count = createCounter(() => location.query.hello);
 
           expect(location.query.hello).toEqual("world");
@@ -210,7 +210,7 @@ describe("Router should", () => {
         const signal = createSignal<LocationChange>({
           value: "/"
         });
-        const { location, navigatorFactory } = createRouterContext(signal);
+        const { location, navigatorFactory } = createRouterContext({ signal });
         const navigate = navigatorFactory();
 
         expect(location.pathname).toBe("/");
@@ -227,7 +227,7 @@ describe("Router should", () => {
         const signal = createSignal<LocationChange>({
           value: "/foo/bar"
         });
-        const { location, navigatorFactory } = createRouterContext(signal);
+        const { location, navigatorFactory } = createRouterContext({ signal });
         const navigate = navigatorFactory();
         const count = createCounter(() => location.pathname);
 
@@ -245,7 +245,7 @@ describe("Router should", () => {
         const signal = createSignal<LocationChange>({
           value: "/"
         });
-        const { navigatorFactory } = createRouterContext(signal);
+        const { navigatorFactory } = createRouterContext({ signal });
         const navigate = navigatorFactory();
         expect(signal[0]().value).toBe("/");
         navigate("/foo/bar");
@@ -262,7 +262,7 @@ describe("Router should", () => {
         const state = { foo: "bar" };
         const signal = createSignal<LocationChange>({ value: "/" });
 
-        const { location, navigatorFactory } = createRouterContext(signal);
+        const { location, navigatorFactory } = createRouterContext({ signal });
         const navigate = navigatorFactory();
 
         expect(location.state).toBeUndefined();
@@ -280,7 +280,7 @@ describe("Router should", () => {
         const state = { foo: "bar" };
         const signal = createSignal<LocationChange>({ value: "/" });
 
-        const { location, navigatorFactory } = createRouterContext(signal);
+        const { location, navigatorFactory } = createRouterContext({ signal });
         const navigate = navigatorFactory();
 
         expect(location.state).toBeUndefined();
@@ -298,7 +298,7 @@ describe("Router should", () => {
         const signal = createSignal<LocationChange>({
           value: "/"
         });
-        const { navigatorFactory } = createRouterContext(signal);
+        const { navigatorFactory } = createRouterContext({ signal });
         const navigate = navigatorFactory();
 
         expect(signal[0]()).toEqual({ value: "/" });
@@ -320,7 +320,7 @@ describe("Router should", () => {
         const signal = createSignal<LocationChange>({
           value: "/"
         });
-        const { navigatorFactory } = createRouterContext(signal);
+        const { navigatorFactory } = createRouterContext({ signal });
         const navigate = navigatorFactory();
         function pushAlot() {
           for (let i = 0; i < 101; i++) {
